+++ index.html
@@ -11,7 +11,7 @@
 
 <body>
   <script src="js/functions.js"></script>
-  <script src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
   <main>
 
     <!-- Фильтрация изображений от других пользователей -->
